requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
cs50==9.3.4
distlib==0.3.8
dnspython==2.7.0
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
